Add progress meter.
We're lying a little about what work's left, since we're counting down issue
numbers but only looking at pull requests.  But since Github doesn't provide
the number of pull requests, it's a decent approximation.  And it'll only be
evident if you have a very uneven distribution of pulls and normal issues.