Update progress meter.
Handling --until required switching the order in which we get issues
from github.  Unfortunately, that means we can no longer estimate our
progress through the pulls (since we're counting forward and have no
idea how many total issues there are, and if using --until it'd be
thrown off anyways).

It's still nice to know that things are happening, though, so we'll now
print out the pull request number we're looking at.  And if you know
your repo, you can get a rough estimate of the progress, so I guess it's
ok.