Always show correct percentage counts
In 2a634cd, I added percentage-calculations to the reports.  However,
this used the total count, which was only being updated on `--basic`.
This didn't throw any exceptions due to the use of a DefaultDict, but we
ended up with all the percentages being 0% when --basic wasn't used.

Now we always record the total number of pulls analyzed.