Fix sometimes crash with --additions and --deletions
`range()` doesn't include it's `stop` value, but we've been acting as if
it did.  This has been ok most of the time, since we bucket values, but
it causes problems if our maximum value is at the end of a bucket and
resides in that bucket by itself.

There are some other places we do a similar sort of `range()`ing, but I
think those all operate on `DefaultDict`s, which means it's ok: it'll
still be the last value added, so the ordering is still correct, and
it'll just default to the default (of zero) when we `+= 1`.